feat(web-app-vercel): NextAuth 設定を lib/auth に集約して認証を共通化
- packages/web-app-vercel/lib/auth.ts を追加し NextAuth の handlers/auth/signIn/ signOut をエクスポート
- API route (/app/api/auth/[...nextauth]/route.ts) を共通 handlers を利用するよう置換
- サインインページ (app/auth/signin/page.tsx) と middleware.ts を共通 signIn/auth に差し替え
- next-auth を v5.0.0-beta に更新
