+++ packages/web-app-vercel/app/auth/signin/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { signIn } from "next-auth/react";
+import { signIn } from "@/lib/auth";
 
 export default function SignIn() {
   return (
