+++ packages/web-app-vercel/middleware.ts
@@ -1,15 +1,12 @@
-export { default } from 'next-auth/middleware';
+import { auth } from "@/lib/auth";
+
+export default auth((req) => {
+    if (!req.auth && !req.nextUrl.pathname.startsWith('/auth')) {
+        const newUrl = new URL('/auth/signin', req.nextUrl.origin);
+        return Response.redirect(newUrl);
+    }
+});
 
 export const config = {
-    matcher: [
-        /*
-         * 以下のパスを除外:
-         * - api (API routes)
-         * - auth (認証ページ)
-         * - _next/static (静的ファイル)
-         * - _next/image (画像最適化)
-         * - favicon.ico (ファビコン)
-         */
-        '/((?!api|auth|_next/static|_next/image|favicon.ico).*)',
-    ],
+    matcher: ['/((?!api|auth|_next/static|_next/image|favicon.ico).*)'],
 };
