+++ packages/web-app-vercel/package.json
@@ -10,7 +10,7 @@
   },
   "dependencies": {
     "next": "15.5.4",
-    "next-auth": "^4.24.10",
+    "next-auth": "^5.0.0-beta.30",
     "next-pwa": "^5.6.0",
     "react": "19.1.0",
     "react-dom": "19.1.0"
